removed legacy search pattern - streamlined more with contact list
